Rename Trellis application test class
